Final Fantasy XII True 720p
It changes handheld mode resolution from 1280x720 to 1600x900 so after downscaling by 80% it gives true 1280x720. But UI is broken thanks to that.
Docked resolution is changed from 1920x1080 to 2020x1350, after downscaling it gives real resolution 1616x1080 (I couldn't bypass arm instructions limits 0x7FF, when 2400 is 0x960.
Best solution would be to find scaler and change it to 100% (or 1.0), but for now I can't find it.